+++ src/quacc/schemas/vasp.py
@@ -7,10 +7,12 @@ from typing import TypeVar
 
 from ase import Atoms
 from emmet.core.tasks import TaskDoc
+from maggma.core import Store
 
 from quacc import SETTINGS
 from quacc.schemas.atoms import atoms_to_metadata
 from quacc.util.atoms import prep_next_run as prep_next_run_
+from quacc.util.db import results_to_db
 from quacc.util.dicts import clean_dict
 from quacc.util.pop_analysis import bader_runner
 
@@ -25,6 +27,7 @@ def summarize_run(
     check_convergence: bool = True,
     remove_empties: bool = False,
     additional_fields: dict | None = None,
+    store: Store | None = None,
 ) -> VaspSchema:
     """
     Get tabulated results from a VASP run and store them in a database-friendly format.
@@ -48,6 +51,8 @@ def summarize_run(
         Whether to remove None values and empty lists/dicts from the TaskDocument.
     additional_fields
         Additional fields to add to the task document.
+    store
+        Maggma Store object to store the results in. If None, `SETTINGS.RESULTS_STORE` will be used.
 
     Returns
     -------
@@ -141,6 +146,7 @@ def summarize_run(
     additional_fields = additional_fields or {}
     run_bader = SETTINGS.VASP_BADER if run_bader is None else run_bader
     dir_path = dir_path or os.getcwd()
+    store = SETTINGS.RESULTS_STORE if store is None else store
 
     # Fetch all tabulated results from VASP outputs files
     # Fortunately, emmet already has a handy function for this
@@ -204,6 +210,13 @@ def summarize_run(
     # makes the structure metadata for us
     atoms_db = atoms_to_metadata(atoms, get_metadata=False, store_pmg=False)
 
-    task_doc = results | atoms_db | additional_fields
+    # Make task document
+    task_doc = clean_dict(
+        results | atoms_db | additional_fields, remove_empties=remove_empties
+    )
 
-    return clean_dict(task_doc, remove_empties=remove_empties)
+    # Store the results
+    if store:
+        results_to_db(store, task_doc)
+
+    return task_doc