+++ tests/recipes/dftb/test_dftb_recipes.py
@@ -10,7 +10,7 @@ from quacc.recipes.dftb.core import relax_job, static_job
 DFTBPLUS_EXISTS = bool(which("dftb+"))
 
 
-def teardown_module():
+def teardown_function():
     for f in os.listdir(os.getcwd()):
         if (
             f.endswith(".log")