+++ tests/schemas/test_ase.py
@@ -11,6 +11,7 @@ from ase.optimize import BFGS
 from ase.thermochemistry import IdealGasThermo
 from ase.units import invcm
 from ase.vibrations import Vibrations
+from maggma.stores import MemoryStore
 from monty.json import MontyDecoder, jsanitize
 
 from quacc.schemas.ase import (
@@ -25,7 +26,7 @@ FILE_DIR = Path(__file__).resolve().parent
 run1 = os.path.join(FILE_DIR, "vasp_run1")
 
 
-def teardown_module():
+def teardown_function():
     for f in os.listdir("."):
         if ".log" in f or ".pckl" in f or ".traj" in f or ".gz" in f:
             os.remove(f)
@@ -45,6 +46,11 @@ def test_summarize_run():
     assert results["atoms"] == atoms
     assert results["results"]["energy"] == atoms.get_potential_energy()
     assert "pull_request" in results["builder_meta"]
+
+    # Test DB
+    store = MemoryStore()
+    summarize_run(atoms, store=store)
+    assert store.count() == 1
 
     # Test remove_empties
     atoms = read(os.path.join(run1, "OUTCAR.gz"))
@@ -107,7 +113,7 @@ def test_summarize_opt_run():
     dyn.run(steps=100)
     traj = read("test.traj", index=":")
 
-    results = summarize_opt_run(dyn, check_convergence=True)
+    results = summarize_opt_run(dyn)
     assert results["nsites"] == len(atoms)
     assert results["atoms"] == traj[-1]
     assert results["results"]["energy"] == atoms.get_potential_energy()
@@ -119,6 +125,29 @@ def test_summarize_opt_run():
     assert "pull_request" in results["builder_meta"]
     assert results["parameters_opt"]["fmax"] == dyn.fmax
     assert results["parameters_opt"]["max_steps"] == 100
+
+    # Test DB
+    atoms = bulk("Cu") * (2, 2, 1)
+    atoms[0].position += [0.1, 0.1, 0.1]
+    atoms.calc = EMT()
+    dyn = BFGS(atoms, trajectory="test.traj")
+    dyn.run(steps=5)
+    traj = read("test.traj", index=":")
+
+    store = MemoryStore()
+    summarize_opt_run(dyn, store=store, check_convergence=False)
+    assert store.count() == 1
+
+    # Test no convergence
+    atoms = bulk("Cu") * (2, 2, 1)
+    atoms[0].position += [0.1, 0.1, 0.1]
+    atoms.calc = EMT()
+    dyn = BFGS(atoms, trajectory="test.traj")
+    dyn.run(steps=5)
+    traj = read("test.traj", index=":")
+
+    with pytest.raises(ValueError):
+        summarize_opt_run(dyn)
 
     # Test remove_empties
     atoms = bulk("Cu") * (2, 2, 1)
@@ -199,6 +228,17 @@ def test_summarize_vib_run():
     assert len(results["results"]["vib_energies"]) == 1
     assert results["results"]["vib_energies"][0] == pytest.approx(0.11507528256667966)
 
+    # Test DB
+    atoms = molecule("N2")
+    atoms.calc = EMT()
+    input_atoms = deepcopy(atoms)
+    vib = Vibrations(atoms)
+    vib.run()
+
+    store = MemoryStore()
+    summarize_vib_run(vib, store=store)
+    assert store.count() == 1
+
     # Test remove_empties
     atoms = molecule("N2")
     atoms.calc = EMT()
@@ -262,6 +302,14 @@ def test_summarize_thermo_run():
     assert results["parameters_thermo"]["vib_freqs"] == [0.34 / invcm]
     assert results["results"]["energy"] == 0
     assert "pull_request" in results["builder_meta"]
+
+    # Test DB
+    atoms = molecule("N2")
+    igt = IdealGasThermo([0.34], "linear", atoms=atoms, spin=0, symmetrynumber=2)
+    summarize_thermo_run(igt)
+    store = MemoryStore()
+    summarize_thermo_run(igt, store=store)
+    assert store.count() == 1
 
     # Test remove_empties
     atoms = molecule("N2")