+++ tests/util/test_db.py
@@ -4,8 +4,7 @@ from shutil import rmtree
 import covalent as ct
 import pytest
 from ase.build import bulk
-from maggma.stores import MemoryStore
-from maggma.stores.mongolike import MontyStore
+from maggma.stores import MemoryStore, MontyStore
 
 from quacc.recipes.emt.core import static_job
 from quacc.util.db import covalent_to_db, results_to_db
@@ -16,7 +15,7 @@ except:
     montydb = None
 
 
-def teardown_module():
+def teardown_function():
     for f in os.listdir(os.getcwd()):
         if "monty.storage" in f:
             os.remove(f)