+++ tests/util/test_sella.py
@@ -16,14 +16,14 @@ except ImportError:
 CWD = os.getcwd()
 
 
-def setup_module():
+def setup_function():
     # Run this test from a fresh directory
     if not os.path.exists("blank_dir"):
         os.mkdir("blank_dir")
     os.chdir("blank_dir")
 
 
-def teardown_module():
+def teardown_function():
     # Clean up
     os.chdir(CWD)
     for f in os.listdir("."):