+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [UNRELEASED]
 
+### Added
+
+- Added a `RESULTS_STORE` quacc setting that, when specified, will automatically store all results in the specified Maggma store.
+
 ### Changed
 
 - The trajectory is now stored in `quacc.schemas.cclib.summarize_run`.